refactor(sitemap-download): move utility functions to utils module
- Create utils.py with archive_file() and format_bytes()
- Remove duplicate code from cli.py
- Update tests to import from utils module
- Update main.py to use new utils module
- Export utils functions from package __init__.py
- Remove unused file:// URL scheme support from models
- Improves separation of concerns (CLI vs utilities)
